Make data population shared
