+++ h-util-ui/Electron/operations/handler.helpers.ts
@@ -1,9 +1,20 @@
+import path from 'path';
+import fs from 'fs/promises';
+
 import { withTimer, promises } from '@common/common';
-import { splitFileNameFromPath } from '@common/fileops';
+import {
+    splitFileNameFromPath,
+    checkSupportedExt,
+    ffMeta,
+    formatDateString,
+    getDateStringForFile,
+} from '@common/fileops';
+import { slugify } from '@shared/common.utils';
 
 import { FileOptions, FileWithMeta, ModuleHandler, ModuleOptions } from '@util/types';
 import { ProcessingModule } from '@shared/common.types';
 import output from '@util/output';
+import { ExtraData, RenameTemplates } from '@shared/common.constants';
 
 type WithFileListHandlingArgs<T extends object> = {
     fileOptions: FileOptions;
@@ -119,3 +130,67 @@ export const addEventLogForReport = (
 export const fileListToFileOptions = (fileList: string[]): FileOptions => ({
     filesWithMeta: fileList.map<FileWithMeta>((filePath) => ({ filePath })),
 });
+
+export type DataDict = Partial<Record<RenameTemplates | ExtraData, string>>;
+
+// todo: refactor to opts obj
+export const populateDataDict = async (dataDict: DataDict, tag: string, filePath: string, mask = 'yy-mm-dd-HH-MM') => {
+    const castTag = tag as RenameTemplates | ExtraData;
+    const { fileName: rawFileName } = splitFileNameFromPath(filePath);
+    const ext = path.extname(rawFileName);
+    const fileName = rawFileName.replace(ext, '');
+
+    switch (castTag) {
+        /** Do all at once so if both are used we only stat once */
+        case RenameTemplates.DateCreated:
+        case RenameTemplates.DateModified:
+        case ExtraData.FileSize:
+            /** Already ran stat - skip */
+            if (dataDict[castTag]) return;
+
+            const stat = await fs.stat(filePath);
+            dataDict[RenameTemplates.DateCreated] = formatDateString(stat.ctime ?? stat.mtime, mask);
+            dataDict[RenameTemplates.DateModified] = formatDateString(stat.mtime ?? stat.ctime, mask);
+            dataDict[ExtraData.FileSize] = String(stat.size);
+            return;
+        case RenameTemplates.ExifTaken:
+            const isImg = checkSupportedExt(filePath, ['img'], true);
+            const { dateStr } = await getDateStringForFile(filePath, isImg, mask);
+            dataDict[RenameTemplates.ExifTaken] = dateStr;
+
+            return;
+        case RenameTemplates.OriginalName:
+            dataDict[castTag] = fileName;
+            return;
+        case RenameTemplates.ParentFolder:
+            const dirPath = path.dirname(filePath);
+            const folderName = path.basename(dirPath);
+            dataDict[RenameTemplates.ParentFolder] = folderName;
+            return;
+        case RenameTemplates.SlugifiedName:
+            dataDict[RenameTemplates.SlugifiedName] = slugify(fileName);
+            return;
+        case RenameTemplates.MetaAlbum:
+        case RenameTemplates.MetaArtist:
+        case RenameTemplates.MetaTitle:
+        case RenameTemplates.MetaTrackNo:
+            if (dataDict[castTag]) return;
+
+            const probeData = await ffMeta.readTags(filePath);
+            const tags = probeData?.format.tags;
+
+            dataDict[RenameTemplates.MetaAlbum] = String(tags?.album) ?? 'UnknownAlbum';
+            dataDict[RenameTemplates.MetaTrackNo] = String(tags?.track) ?? '0';
+            dataDict[RenameTemplates.MetaTitle] = String(tags?.title) ?? 'UnknownTitle';
+            dataDict[RenameTemplates.MetaArtist] = String(tags?.artist) ?? 'UnknownArtist';
+            return;
+        case ExtraData.Extension:
+            dataDict[ExtraData.Extension] = path.extname(fileName);
+            return;
+        case ExtraData.ocr:
+            // todo: pass in param
+            return;
+        default:
+            return;
+    }
+};
