+++ h-util-ui/Electron/operations/modules/dynamicRename.handler.ts
@@ -12,8 +12,7 @@ import {
 import { defaultTimeMask, RenameTemplates } from '@shared/common.constants';
 import { ConfigError } from '@util/errors';
 import { ModuleHandler, ModuleOptions } from '@util/types';
-import { slugify } from '@shared/common.utils';
-import { addEventLogForReport } from '../handler.helpers';
+import { addEventLogForReport, DataDict, populateDataDict } from '../handler.helpers';
 
 type RequiredDataContext = {
     requiredData?: Set<DataNeeded>;
@@ -21,8 +20,6 @@ type RequiredDataContext = {
     tags?: string[];
     testMode?: boolean;
 };
-
-type DataDict = Partial<Record<RenameTemplates, string>>;
 
 const dynamicRenameHandler: ModuleHandler<RequiredDataContext> = {
     handler: async (fileWithMeta, opts) => {
@@ -62,62 +59,9 @@ const dynamicRenameHandler: ModuleHandler<RequiredDataContext> = {
 
 export default dynamicRenameHandler;
 
-const populateDataDict = async (dataDict: DataDict, tag: string, filePath: string, mask = 'yy-mm-dd-HH-MM') => {
-    const castTag = tag as RenameTemplates;
-    const { fileName: rawFileName } = splitFileNameFromPath(filePath);
-    const ext = path.extname(rawFileName);
-    const fileName = rawFileName.replace(ext, '');
-
-    switch (castTag) {
-        /** Do both at once so if both are used we only stat once */
-        case RenameTemplates.DateCreated:
-        case RenameTemplates.DateModified:
-            /** Already ran stat - skip */
-            if (dataDict[castTag]) return;
-
-            const stat = await fs.stat(filePath);
-            dataDict[RenameTemplates.DateCreated] = formatDateString(stat.ctime ?? stat.mtime, mask);
-            dataDict[RenameTemplates.DateModified] = formatDateString(stat.mtime ?? stat.ctime, mask);
-
-            return;
-        case RenameTemplates.ExifTaken:
-            const isImg = checkSupportedExt(filePath, ['img'], true);
-            const { dateStr } = await getDateStringForFile(filePath, isImg, mask);
-            dataDict[RenameTemplates.ExifTaken] = dateStr;
-
-            return;
-        case RenameTemplates.OriginalName:
-            dataDict[castTag] = fileName;
-            return;
-        case RenameTemplates.ParentFolder:
-            const dirPath = path.dirname(filePath);
-            const folderName = path.basename(dirPath);
-            dataDict[RenameTemplates.ParentFolder] = folderName;
-            return;
-        case RenameTemplates.SlugifiedName:
-            dataDict[RenameTemplates.SlugifiedName] = slugify(fileName);
-            return;
-        case RenameTemplates.MetaAlbum:
-        case RenameTemplates.MetaArtist:
-        case RenameTemplates.MetaTitle:
-        case RenameTemplates.MetaTrackNo:
-            if (dataDict[castTag]) return;
-
-            const probeData = await ffMeta.readTags(filePath);
-            const tags = probeData?.format.tags;
-
-            dataDict[RenameTemplates.MetaAlbum] = String(tags?.album) ?? 'UnknownAlbum';
-            dataDict[RenameTemplates.MetaTrackNo] = String(tags?.track) ?? '0';
-            dataDict[RenameTemplates.MetaTitle] = String(tags?.title) ?? 'UnknownTitle';
-            dataDict[RenameTemplates.MetaArtist] = String(tags?.artist) ?? 'UnknownArtist';
-            return;
-        default:
-            return;
-    }
-};
-
 /**
  * Extract all info if first run
+ * This informs what data we need to extract from file based on tags
  */
 const populateContext = (stringTemplate: string, opts: Partial<ModuleOptions<RequiredDataContext>>) => {
     if (!opts.context) opts.context = {};
